refactor(direct-control): extract destination analysis source
Move the destination analysis embedded App UI source into src/play/tactical/destination.ts while keeping the public wrapper and command builder in the facade. Keep its battlefield-source dependency explicit because destination analysis composes the battlefield read policy.

Record the tactical/progression read atom set as extracted in the support modularization OpenSpec workstream. This is source ownership proof only and does not claim live-game runtime proof.

Validation: bun run --cwd packages/civ7-direct-control test -- tactical-reads.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
