+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -122,7 +122,7 @@ runtime/direct-control claims.
   - [x] 4.5.3 Extract ready-city embedded source owner while keeping the public
         wrapper in the facade.
 - [ ] 4.6 Extract operation validation/send/postcondition atoms.
-- [ ] 4.7 Extract settlement/tactical/progression read atoms.
+- [x] 4.7 Extract settlement/tactical/progression read atoms.
   - [x] 4.7.1 Extract settlement recommendation embedded source owner while
         keeping the public wrapper in the facade.
   - [x] 4.7.2 Extract traditions view embedded source owner while keeping the
@@ -133,6 +133,8 @@ runtime/direct-control claims.
         public wrapper in the facade.
   - [x] 4.7.5 Extract battlefield scan embedded source owner while keeping the
         public wrapper in the facade.
+  - [x] 4.7.6 Extract destination analysis embedded source owner while keeping
+        the public wrapper in the facade.
 - [ ] 4.8 Export stable types/constants only after module owners are defined.
   - [x] 4.8.1 Extract ComponentID primitive and direct-control error owner
         modules behind the existing package facade.
